+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Scene buttons generate actors in folders ([#237](https://github.com/ben/foundry-ironsworn/pull/237))
 - …and also drop a token in the center of the viewport ([#239](https://github.com/ben/foundry-ironsworn/pull/239))
+- Wire up settlement/star oracles in location sheet ([#240](https://github.com/ben/foundry-ironsworn/pull/240))
 
 ## 1.10.25